refactor(angular): move deocrators.ts to preview root
Preview/angular/* is now legacy. The decorators defined here are not. So they are moved to preview root
