+++ app/angular/src/client/index.ts
@@ -13,7 +13,7 @@ export * from './preview/types-6-0';
 
 export { StoryFnAngularReturnType as IStory } from './preview/types';
 
-export { moduleMetadata, componentWrapperDecorator } from './preview/angular/decorators';
+export { moduleMetadata, componentWrapperDecorator } from './preview/decorators';
 
 if (module && module.hot && module.hot.decline) {
   module.hot.decline();
