+++ app/angular/src/client/preview/decorators.test.ts
@@ -2,7 +2,7 @@ import addons, { mockChannel } from '@storybook/addons';
 
 import { Component } from '@angular/core';
 import { moduleMetadata } from './decorators';
-import { addDecorator, storiesOf, clearDecorators, getStorybook } from '..';
+import { addDecorator, storiesOf, clearDecorators, getStorybook } from '.';
 
 class MockModule {}
 class MockModuleTwo {}
